Fix to work with Filter plugin
