+++ examples/reemit_filter.conf
@@ -1,0 +1,23 @@
+<source>
+  type gc_stat
+  emit_interval 1s
+  tag gc_stat
+</source>
+
+<match **>
+  @type copy
+  <store>
+    @type stdout
+  </store>
+  <store>
+    @type reemit
+  </store>
+</match>
+
+<filter **>
+  @type stdout
+</filter>
+
+<match **>
+  @type stdout
+</match>
